requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
twilio==8.10.1
python-dotenv==1.0.0
geopy==2.4.1
requests==2.31.0
PyJWT==2.8.0
bcrypt==4.1.1
python-dateutil==2.8.2

